requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
docker==7.1.0
gitdb==4.0.12
GitPython==3.1.44
GPUtil==1.4.0
humanfriendly==10.0
idna==3.10
psutil==6.1.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-on-whales==0.74.0
PyYAML==6.0.2
pyyml==0.0.2
requests==2.32.3
setuptools==75.7.0
smmap==5.0.2
typing_extensions==4.12.2
urllib3==2.3.0
urwid==2.6.16
vulkan==1.3.275.1
wcwidth==0.2.13
